+++ agent_api_http/src/v0/authorization/mod.rs
@@ -3,7 +3,7 @@ pub mod authorization_server;
 
 pub mod error;
 
-use crate::authorization::authorization_server::consent::{get_consent, post_consent};
+use crate::v0::authorization::authorization_server::consent::{get_consent, post_consent};
 use crate::API_VERSION;
 use agent_authorization::state::AuthorizationState;
 use agent_issuance::state::IssuanceState;
