+++ agent_api_http/src/v0/mod.rs
@@ -1,0 +1,7 @@
+pub mod authorization;
+pub mod holder;
+pub mod identity;
+pub mod issuance;
+pub mod library;
+pub mod templates;
+pub mod verification;
